+++ knowledge/nel/__init__.py
@@ -1,7 +1,9 @@
 # -*- coding: utf-8 -*-
-# Copyright © 2021 Wacom. All rights reserved.
+# Copyright © 2021-23 Wacom. All rights reserved.
 """"
 Named Entity Linking
+--------------------
+This module provides the Named Entity Linking (NEL) functionality on top of the knowledge graph entities.
 """
 
 __all__ = ['base', 'engine']